unicorn_properties/pyproject.toml
[tool.poetry]
name = "properties_service"
version = "0.2.0"
description = "Unicorn Properties Property Service"
authors = ["Amazon Web Services"]
packages = [
    { include = "properties_service", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.13"
boto3 = "^1.35"
aws-lambda-powertools = { extras = ["aws-sdk"], version = "^2.23.0" }
aws-xray-sdk = "^2.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
requests = "^2.31.0"
moto = "^4.1.13"
importlib-metadata = "^6.8.0"
pyyaml = "^6.0.1"
arnparse = "^0.0.2"
aws-cdk-lib = "^2.167.0"
constructs = ">=10.0.0,<11.0.0"
cdk-nag = "^2.33.1"
unicorn-shared = {path = "../unicorn_shared", develop = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -vv -W ignore::UserWarning"
testpaths = ["tests/unit"]

[tool.ruff]
line-length = 150
